Move RadioGroup files in Radio folder
