+++ src/components/styles.scss
@@ -13,7 +13,7 @@
 @import './ListGroup/ListGroup';
 @import './Mask/Mask';
 @import './Radio/Radio';
-@import './RadioGroup/RadioGroup';
+@import './Radio/RadioGroup';
 @import './Space/Space';
 @import './Switch/Switch';
 @import './Tooltip/Tooltip';
